Backend: Secure API endpoints with permission checks
* WIP ExecutionController
